refactor(tests): split utils out of smoke.test.js
